8-Agentic AI and Autonomous AI Systems/module03_1/requirements.txt
langchain
langchain-google-genai
fastapi
uvicorn
langserve
sse_starlette
python-dotenv
